% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{ExecutionError}
\alias{ExecutionError}
\title{ExecutionError Object}
\usage{
ExecutionError(scriptStackTraceElements = NULL, errorMessage = NULL,
  errorType = NULL)
}
\arguments{
\item{scriptStackTraceElements}{An array of objects that provide a stack trace through the script to show where the execution failed, with the deepest call first}

\item{errorMessage}{The error message thrown by Apps Script, usually localized into the user's language}

\item{errorType}{The error type, for example `TypeError` or `ReferenceError`}
}
\value{
ExecutionError object
}
\description{
ExecutionError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object that provides information about the nature of an error in the Apps Script Execution API. If an `run` call succeeds but the script function (or Apps Script itself) throws an exception, the response body's `error` field will contain a `Status` object. The `Status` object's `details` field will contain an array with a single one of these `ExecutionError` objects.
}

