% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{ExecutionRequest}
\alias{ExecutionRequest}
\title{Google Apps Script Execution API Objects 
Executes Google Apps Script projects.}
\usage{
ExecutionRequest(function. = NULL, parameters = NULL, sessionState = NULL,
  devMode = NULL)
}
\arguments{
\item{function.}{The name of the function to execute in the given script}

\item{parameters}{The parameters to be passed to the function being executed}

\item{sessionState}{This field is not used}

\item{devMode}{If `true` and the user is an owner of the script, the script runs at the most recently saved version rather than the version deployed for use with the Execution API}
}
\value{
ExecutionRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:22:54
filename: /Users/mark/dev/R/autoGoogleAPI/googlescriptv1.auto/R/script_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ExecutionRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to run the function in a script. The script is identified by the specified `script_id`. Executing a function on a script will return results based on the implementation of the script.
}
\seealso{
Other ExecutionRequest functions: \code{\link{.run}}
}

