% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{ExecutionResponse}
\alias{ExecutionResponse}
\title{ExecutionResponse Object}
\usage{
ExecutionResponse(result = NULL, status = NULL)
}
\arguments{
\item{result}{The return value of the script function}

\item{status}{No description}
}
\value{
ExecutionResponse object
}
\description{
ExecutionResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object that provides the return value of a function executed through the Apps Script Execution API. If an `run` call succeeds and the script function returns successfully, the response body's `response` field will contain this `ExecutionResponse` object.
}

