% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(Operation.response = NULL, Operation.metadata = NULL,
  done = NULL, response = NULL, name = NULL, error = NULL,
  metadata = NULL)
}
\arguments{
\item{Operation.response}{The \link{Operation.response} object or list of objects}

\item{Operation.metadata}{The \link{Operation.metadata} object or list of objects}

\item{done}{This field is not used}

\item{response}{If the script function returns successfully, this field will contain an `ExecutionResponse` object with the function's return value as the object's `result` field}

\item{name}{This field is not used}

\item{error}{If a `run` call succeeds but the script function (or Apps Script itself) throws an exception, this field will contain a `Status` object}

\item{metadata}{This field is not used}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response will not arrive until the function finishes executing. The maximum runtime is listed in the guide to [limitations in Apps Script](https://developers.google.com/apps-script/guides/services/quotas#current_limitations).<p>If the script function returns successfully, the `response` field will contain an `ExecutionResponse` object with the function's return value in the object's `result` field.</p><p>If the script function (or Apps Script itself) throws an exception, the `error` field will contain a `Status` object. The `Status` object's `details` field will contain an array with a single `ExecutionError` object that provides information about the nature of the error.</p><p>If the `run` call itself fails (for example, because of a malformed request or an authorization error), the method will return an HTTP response code in the 4XX range with a different format for the response body. Client libraries will automatically convert a 4XX response into an exception class.</p>
}
\seealso{
Other Operation functions: \code{\link{Operation.metadata}},
  \code{\link{Operation.response}}
}
