% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{Operation.response}
\alias{Operation.response}
\title{Operation.response Object}
\usage{
Operation.response()
}
\value{
Operation.response object
}
\description{
Operation.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If the script function returns successfully, this field will contain an `ExecutionResponse` object with the function's return value as the object's `result` field.
}
\seealso{
Other Operation functions: \code{\link{Operation.metadata}},
  \code{\link{Operation}}
}

