% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{ScriptStackTraceElement}
\alias{ScriptStackTraceElement}
\title{ScriptStackTraceElement Object}
\usage{
ScriptStackTraceElement(lineNumber = NULL, function. = NULL)
}
\arguments{
\item{lineNumber}{The line number where the script failed}

\item{function.}{The name of the function that failed}
}
\value{
ScriptStackTraceElement object
}
\description{
ScriptStackTraceElement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A stack trace through the script that shows where the execution failed.
}
