% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{Status}
\alias{Status}
\title{Status Object}
\usage{
Status(code = NULL, message = NULL, details = NULL)
}
\arguments{
\item{code}{The status code, which should be an enum value of google}

\item{message}{A developer-facing error message, which should be in English}

\item{details}{An array that contains a single `ExecutionError` object that provides information about the nature of the error}
}
\value{
Status object
}
\description{
Status Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If a `run` call succeeds but the script function (or Apps Script itself) throws an exception, the response body's `error` field will contain this `Status` object.
}

