% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{ExecutionError}
\alias{ExecutionError}
\title{ExecutionError Object}
\usage{
ExecutionError(errorType = NULL, errorMessage = NULL,
  scriptStackTraceElements = NULL)
}
\arguments{
\item{errorType}{The error type, for example `TypeError` or `ReferenceError`}

\item{errorMessage}{The error message thrown by Apps Script, usually localized into the user's}

\item{scriptStackTraceElements}{An array of objects that provide a stack trace through the script to show}
}
\value{
ExecutionError object
}
\description{
ExecutionError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object that provides information about the nature of an error in the AppsScript Execution API. If an`run` call succeeds but thescript function (or Apps Script itself) throws an exception, the responsebody's `error` field contains a`Status` object. The `Status` object's `details` fieldcontains an array with a single one of these `ExecutionError` objects.
}
