% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{ExecutionRequest}
\alias{ExecutionRequest}
\title{ExecutionRequest Object}
\usage{
ExecutionRequest(function. = NULL, devMode = NULL, parameters = NULL,
  sessionState = NULL)
}
\arguments{
\item{function.}{The name of the function to execute in the given script}

\item{devMode}{If `true` and the user is an owner of the script, the script runs at the}

\item{parameters}{The parameters to be passed to the function being executed}

\item{sessionState}{For Android add-ons only}
}
\value{
ExecutionRequest object
}
\description{
ExecutionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to run the function in a script. The script is identified by thespecified `script_id`. Executing a function on a script returns resultsbased on the implementation of the script.
}
\seealso{
Other ExecutionRequest functions: \code{\link{.run}}
}
