% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{ExecutionResponse}
\alias{ExecutionResponse}
\title{ExecutionResponse Object}
\usage{
ExecutionResponse(result = NULL)
}
\arguments{
\item{result}{The return value of the script function}
}
\value{
ExecutionResponse object
}
\description{
ExecutionResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object that provides the return value of a function executed through theApps Script Execution API. If a`run` call succeeds and thescript function returns successfully, the response body's`response` field contains this`ExecutionResponse` object.
}
