% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_objects.R
\name{Status}
\alias{Status}
\title{Google Apps Script Execution API Objects 
Executes Google Apps Script projects.}
\usage{
Status(details = NULL, code = NULL, message = NULL)
}
\arguments{
\item{details}{An array that contains a single `ExecutionError` object that provides information about the nature of the error}

\item{code}{The status code}

\item{message}{A developer-facing error message, which is in English}
}
\value{
Status object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:11:39
filename: /Users/mark/dev/R/autoGoogleAPI/googlescriptv1.auto/R/script_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Status Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If a `run` call succeeds but the script function (or Apps Script itself) throws an exception, the response body's `error` field will contain this `Status` object.
}
